Cleanup engine/engineobj code.
engines/pandas/engine.py is a misnamed file in an incorrect directory.
That file is really only a Parquet reader code. As such it needs to be
called that and made to reside in a directory called DB or some such thing.
So, moved a lot of the unnecessary semantics of analysis semantics that
were present in this file. All that semantic information belongs in the
engineobj.py file which is really the basic analysis object.

The engine.py code needs only to (i) know what table to read, (ii) what
columns to read, (iii) what predicate pushdown to apply, and (iv) dedup
rows based on the specified time/view fields. Everything else does not
belong in this file.

This is exactly what this patch achieves.